Set ARC property semantics to "copy" for block.
It seems that ARC should be copying automatically when assigning to a "strong" property of a block type, and this generally appears to be working. But there also appears to be a bug in certain compiler versions in which release mode contains optimizations which do not copy the block as expected (thus leading to an EXC_BAD_ACCESS.) See http://stackoverflow.com/questions/12061756/blocks-and-arc-copy-or-crash-with-release-build-caused-by-optimization-level for more information.